reduce layout shift from dynamic vault
