+++ src/components/Layout.tsx
@@ -1,25 +1,18 @@
-import dynamic from "next/dynamic"
 import Image from "next/image"
 import Link from "next/link"
 import { FC, PropsWithChildren } from "react"
 
+import ConnectWalletButton from "@/components/ConnectWallet/ConnectWalletButton"
 import ConnectWalletModal, {
   DisconnectWalletModal,
 } from "@/components/ConnectWallet/ConnectWalletModal"
 import ExternalLinks from "@/components/ExternalLinks"
+import NetworkSelector from "@/components/NetworkSelector"
 
 import { useConnectWallet } from "@/store/connectWallet"
 
 import FortressBackground from "~/images/fortress-background.gif"
 import FortressLogo from "~/svg/fortress-logo.svg"
-
-const ConnectWalletButton = dynamic(
-  () => import("@/components/ConnectWallet/ConnectWalletButton"),
-  { ssr: false }
-)
-const NetworkSelector = dynamic(() => import("@/components/NetworkSelector"), {
-  ssr: false,
-})
 
 const Layout: FC<PropsWithChildren> = ({ children }) => {
   const [connectModal, setConnectModal] = useConnectWallet((state) => [
