+++ src/components/Vault/VaultTable.tsx
@@ -1,19 +1,19 @@
+import dynamic from "next/dynamic"
 import { FC } from "react"
 
-import { useVaultAddresses } from "@/hooks/data"
 import { VaultProps } from "@/hooks/types"
 
-import Spinner from "@/components/Spinner"
-import VaultRow from "@/components/Vault/VaultRow"
 import {
   VaultTableHeader,
   VaultTableRow,
 } from "@/components/Vault/VaultTableNode"
 
+const VaultTableBody = dynamic(
+  () => import("@/components/Vault/VaultTableBody"),
+  { ssr: false }
+)
+
 const VaultTable: FC<Pick<VaultProps, "type">> = ({ type }) => {
-  const { data: vaultAddresses, isLoading } = useVaultAddresses({
-    type,
-  })
   return (
     <div className="" role="table">
       {/* Table headings */}
@@ -30,40 +30,9 @@ const VaultTable: FC<Pick<VaultProps, "type">> = ({ type }) => {
       </div>
 
       {/* Table body */}
-      <div className="space-y-2" role="rowgroup">
-        {isLoading ? (
-          <VaultsLoading />
-        ) : !vaultAddresses?.length ? (
-          <NoVaultsFound />
-        ) : (
-          vaultAddresses?.map((address, i) => (
-            <VaultRow key={`pool-${i}`} asset={address} type={type} />
-          ))
-        )}
-      </div>
+      <VaultTableBody type={type} />
     </div>
   )
 }
 
 export default VaultTable
-
-const VaultsLoading: FC = () => {
-  return (
-    <VaultTableRow
-      className="flex h-44 items-center rounded-t-none"
-      aria-label="Loading vaults..."
-    >
-      <span className="col-span-full text-center" aria-hidden="true">
-        <Spinner className="h-10 w-10" />
-      </span>
-    </VaultTableRow>
-  )
-}
-
-const NoVaultsFound: FC = () => {
-  return (
-    <VaultTableRow className="flex h-44 items-center rounded-t-none">
-      <h2 className="col-span-full text-center text-2xl">No vaults found...</h2>
-    </VaultTableRow>
-  )
-}
