+++ src/pages/_app.tsx
@@ -1,7 +1,12 @@
 import { Inter, VT323 } from "@next/font/google"
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
+import {
+  Hydrate,
+  QueryClient,
+  QueryClientProvider,
+} from "@tanstack/react-query"
 import { Analytics } from "@vercel/analytics/react"
 import { AppProps } from "next/app"
+import { useState } from "react"
 
 import "@/styles/globals.css"
 
@@ -15,15 +20,17 @@ const vt323 = VT323({
 })
 
 function MyApp({ Component, pageProps }: AppProps) {
-  const queryClient = new QueryClient({
-    defaultOptions: {
-      queries: {
-        refetchOnMount: false,
-        refetchOnWindowFocus: false,
-        retryOnMount: false,
+  const [queryClient] = useState(
+    new QueryClient({
+      defaultOptions: {
+        queries: {
+          refetchOnMount: false,
+          refetchOnWindowFocus: false,
+          retryOnMount: false,
+        },
       },
-    },
-  })
+    })
+  )
 
   return (
     <>
@@ -36,9 +43,11 @@ function MyApp({ Component, pageProps }: AppProps) {
 
       <div className="font-sans">
         <QueryClientProvider client={queryClient}>
-          <WagmiProvider>
-            <Component {...pageProps} />
-          </WagmiProvider>
+          <Hydrate state={pageProps.dehydratedState}>
+            <WagmiProvider>
+              <Component {...pageProps} />
+            </WagmiProvider>
+          </Hydrate>
         </QueryClientProvider>
       </div>
